requirements-prd.txt
dispatch==0.5.6
git+git://github.com/ubyssey/django-google-storage.git
django-gae-backends
django-canonical-domain
mysql-python
ua-parser
user-agents
django-user-agents
google-cloud-core==0.27.1
google-cloud-datastore==1.3.0
google-cloud-storage==1.4.0
